fix(ws6-prep): clear all clippy -D warnings in raylib-sys and raylib (Task 5b)
Zero warnings on all four clippy invocations:
  cargo clippy -p raylib-sys -- -D warnings
  cargo clippy -p raylib-sys --no-default-features --features software_renderer -- -D warnings
  cargo clippy -p raylib --lib --features SUPPORT_MODULE_R{SHAPES,TEXTURES,TEXT,MODELS,AUDIO} -- -D warnings
  cargo clippy -p raylib --lib --no-default-features --features software_renderer,... -- -D warnings

Fixes applied (idiom-only, no behaviour changes):
- useless_conversion: drop redundant .into() on same-type FFI returns (color, math, rect,
  vector types) across color.rs, camera.rs, collision.rs, drawing.rs, input.rs, misc.rs,
  models.rs, shaders.rs, text.rs, texture.rs, window.rs
- needless_return: remove trailing `return` statements across audio.rs, callbacks.rs,
  collision.rs, ease.rs, text.rs
- needless_lifetimes / extra_unused_lifetimes: use '_ anonymous lifetime in impl blocks
  across audio.rs, automation.rs, callbacks.rs, drawing.rs, file.rs, misc.rs, models.rs,
  rlgl/immediate.rs, rlgl/matrix.rs
- needless_borrow: &self.0 → self.0 in Deref impls (drawing.rs, misc.rs)
- missing_safety_doc: add # Safety sections to unsafe fns via macro edits in macros.rs
  (impl_wrapper!/gen_from_raw_wrapper!) + explicit docs in audio.rs, callbacks.rs,
  models.rs, shaders.rs, texture.rs, window.rs
- double_must_use: remove bare #[must_use] from Result-returning functions (audio.rs,
  models.rs, text.rs, texture.rs, window.rs)
- too_many_arguments: #[allow] with comment on update_model_animation_ex, draw_ring,
  draw_ring_lines, draw_text_pro, draw_billboard_pro
- ptr_arg: &Vec<u8> → &[u8] in new_music_from_memory (audio.rs)
- unnecessary_cast: remove same-type casts in collision.rs, drawing.rs, input.rs,
  misc.rs, mod.rs, models.rs
- legacy_numeric_constants: remove `use std::usize` (misc.rs)
- suspicious_doc_comments: ///! → //! (logging.rs)
- manual_map: match → .copied() in RandomSequence iterator (misc.rs)
- transmute_ptr_to_ref / missing_transmute_annotations: safe pointer casts in
  camera.rs, collision.rs, input.rs, models.rs
- derivable_impls: #[allow] with comment on WeakTexture2D::default (macro-generated
  struct; cannot use #[derive]) and TraceLogLevel (LOG_INFO ≠ discriminant 0)
- needless_doctest_main: remove fn main() wrappers in window.rs and lib.rs doctests
- type_complexity: extract AudioCallbackFn type alias (callbacks/audio_stream_callback.rs)
- redundant_field_names: shorthand init in stream_processor_with_user_data_wrapper.rs
- unused_unit: remove -> () from FnMut bounds (callbacks.rs)
- unused_imports: remove unused `null_mut` import (texture.rs)

All 18 doctests pass. Build clean with SUPPORT_MODULE_RAUDIO + SUPPORT_MODULE_RTEXTURES.

Attribution: consolidates idiom fixes originally proposed by AmityWilder in PR #224.

+++ raylib/src/core/mod.rs
@@ -327,7 +327,7 @@ impl RaylibBuilder {
         }
 
         unsafe {
-            ffi::SetConfigFlags(flags as u32);
+            ffi::SetConfigFlags(flags);
         }
 
         unsafe {
